media-libs/mesa: Only remove files if there are files to remove
Noticed by Arfrever.

+++ media-libs/mesa/mesa-18.1.1-r1.ebuild
@@ -401,11 +401,13 @@ multilib_src_configure() {
 multilib_src_install() {
 	emake install DESTDIR="${D}"
 
-	# These files are now provided by >=dev-libs/wayland-1.15.0
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so" || die
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1" || die
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1.0.0" || die
-	rm "${ED}/usr/$(get_libdir)/pkgconfig/wayland-egl.pc" || die
+	if use wayland; then
+		# These files are now provided by >=dev-libs/wayland-1.15.0
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so" || die
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1" || die
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1.0.0" || die
+		rm "${ED}/usr/$(get_libdir)/pkgconfig/wayland-egl.pc" || die
+	fi
 
 	if use opencl; then
 		ebegin "Moving Gallium/Clover OpenCL implementation for dynamic switching"
